Fixed controls in VR
