+++ src/main/java/nl/tudelft/contextproject/model/entities/control/PlayerControl.java
@@ -3,8 +3,10 @@ package nl.tudelft.contextproject.model.entities.control;
 import com.jme3.bullet.control.CharacterControl;
 import com.jme3.input.controls.ActionListener;
 import com.jme3.math.Vector3f;
+import com.jme3.renderer.Camera;
 import com.jme3.scene.Spatial;
 
+import jmevr.app.VRApplication;
 import nl.tudelft.contextproject.Main;
 import nl.tudelft.contextproject.model.entities.Entity;
 import nl.tudelft.contextproject.model.entities.VRPlayer;
@@ -29,11 +31,11 @@ public class PlayerControl implements EntityControl, ActionListener {
 
 	@Override
 	public void move(float tpf) {
-		//TODO this will change after VR support is implemented
-		Vector3f camDir = Main.getInstance().getCamera().getDirection();
-		Vector3f camLeft = Main.getInstance().getCamera().getLeft();
+		Camera camera = VRApplication.getVRViewManager().getCamLeft();
+		Vector3f camDir = camera.getDirection().mult(2.0f);
+		Vector3f camLeft = camera.getLeft().mult(2.0f);
 		walkDirection = new Vector3f();
-
+		System.out.println(camera);
 		if (left) {
 			walkDirection.addLocal(camLeft.normalizeLocal().multLocal(SIDE_WAY_SPEED_MULTIPLIER));
 		}
@@ -46,7 +48,7 @@ public class PlayerControl implements EntityControl, ActionListener {
 		if (down) {
 			walkDirection.addLocal(new Vector3f(-camDir.getX(), 0, -camDir.getZ()).normalizeLocal().multLocal(STRAIGHT_SPEED_MULTIPLIER));
 		}
-
+		
 		playerControl.setWalkDirection(walkDirection);
 		spatial.setLocalTranslation(playerControl.getPhysicsLocation().add(0, -2, 0));
 	}
